@@ -5,15 +5,15 @@
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
-    "@meilisearch/instant-meilisearch": "0.8.0",
-    "instantsearch.css": "7.4.2",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-instantsearch-dom": "6.12.1",
-    "react-scripts": "3.0.1"
+    "@meilisearch/instant-meilisearch": "0.9.0",
+    "instantsearch.css": "7.4.5",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-instantsearch-dom": "6.36.0",
+    "react-scripts": "5.0.1"
   },
   "devDependencies": {
-    "typescript": "3.8.3"
+    "typescript": "4.8.4"
   },
   "scripts": {
     "start": "react-scripts start",
